Improved the quality of PR
